+++ client/src/App.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 
 import './App.css';
 
-import Layout from './views/Layout';
-import Dashboard from './views/Dashboard';
+import Layout from './components/Layout/Layout';
+import Dashboard from './components/Dashboard/Dashboard';
 
 function App() {
   return (
