+++ client/src/components/Auth/PrivateRoute.js
@@ -1,6 +1,6 @@
 import React, { useEffect } from "react";
 import { Route } from "react-router-dom";
-import { useAuth0 } from "../Auth/auth0-wrapper";
+import { useAuth0 } from "./auth0-wrapper";
 
 const PrivateRoute = ({ component: Component, path, ...rest }) => {
   const { isAuthenticated, loginWithRedirect } = useAuth0();
