+++ client/src/components/Auth/auth0-wrapper.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect, useContext } from 'react';
 import createAuth0Client from '@auth0/auth0-spa-js';
 import { useStateValue } from 'react-conflux';
-import { globalContext } from '../store/reducers/globalReducer';
+import { globalContext } from '../../store/reducers/globalReducer';
 import { register } from './auth-queries';
 
 const DEFAULT_REDIRECT_CALLBACK = () =>
