+++ client/src/components/OnBoarding/OnboardForm.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { useStateValue } from 'react-conflux';
-import { addSchoolDetails } from './OnBoardQueries';
+import { addSchoolDetails } from './onboardQueries';
 import {
   globalContext,
   ON_BOARD_DETAILS
